requirements.txt
Flask==2.0.3
tqdm==4.62.3
google-cloud-documentai==2.0.0
google-auth==1.35.0
PyPDF2==1.26.0
tika==1.24
pandas==1.3.3
python-dotenv==0.19.2
